fix(tts): cast pyopenjtalk float64 waveform to int16
pyopenjtalk returns float64 samples in int16 amplitude range. SounddeviceAudioOutput.to_float32() only normalizes from int16, so passing the raw float64 array would feed values like ±13000 into a float32 stream expecting [-1.0, 1.0], producing severe clipping.

Cast to int16 so the existing audio output conversion path normalizes correctly. Test now asserts dtype and a representative sample value.

+++ dimos/stream/audio/tts/node_open_jtalk.py
@@ -16,6 +16,7 @@
 import threading
 import time
 
+import numpy as np
 import pyopenjtalk  # type: ignore[import-not-found]
 from reactivex import Observable, Subject
 
@@ -82,8 +83,10 @@ class OpenJTalkTTSNode(AbstractTextConsumer, AbstractAudioEmitter, AbstractTextE
         try:
             waveform, _ = pyopenjtalk.tts(text)
             self.text_subject.on_next(text)
+            # pyopenjtalk returns float64 in int16 amplitude range;
+            # cast to int16 so SounddeviceAudioOutput's to_float32 normalizes correctly.
             audio_event = AudioEvent(
-                data=waveform,
+                data=waveform.astype(np.int16),
                 sample_rate=SAMPLE_RATE,
                 timestamp=time.time(),
                 channels=1,
